First pieces of the Extended XMLRPC API for Wordpress
This commit adds the `exapi_getRecentPosts()' method. It is a binding to
the original `wp_get_recent_posts()' but different from the two
alternatives `{blogger,mw}_getRecentPosts()', it is possible to pass all
parameters supported by the query parameter.

So, you can for example, from python, call this method as the following
examples:

  >>> exapi.getRecentPosts(post_status='published')
  >>> exapi.getRecentPosts(numposts=2, category_name='Uncategorized')