A heavy refactoring in the getRecentPosts() method
This method was divided in a lot of small functions that cares specifically
about one or a few attributes of the post. These small helpers were placed
in the `exapi.post.php' file.

Some things still hardcoded or broken, like getting other sizes of
thumbnails, but it's gonna be fixed in the next few commits.